Conflicts:
	src/main/java/com/cdkj/loan/ao/IBudgetOrderAO.java
	src/main/java/com/cdkj/loan/ao/impl/BudgetOrderAOImpl.java
	src/main/java/com/cdkj/loan/bo/IBudgetOrderBO.java
	src/main/java/com/cdkj/loan/bo/impl/BudgetOrderBOImpl.java
	src/main/java/com/cdkj/loan/dao/IBudgetOrderDAO.java
	src/main/java/com/cdkj/loan/dao/impl/BudgetOrderDAOImpl.java
	src/main/resources/mybatis/BudgetOrderMapper.xml

+++ src/main/java/com/cdkj/loan/domain/CreditUser.java
@@ -115,7 +115,7 @@ public class CreditUser extends ABaseDO {
     private Long monthIncome;
 
     // 结息
-    private Double settleInterest;
+    private Long settleInterest;
 
     // 余额
     private Long balance;
@@ -390,11 +390,11 @@ public class CreditUser extends ABaseDO {
         this.monthIncome = monthIncome;
     }
 
-    public Double getSettleInterest() {
+    public Long getSettleInterest() {
         return settleInterest;
     }
 
-    public void setSettleInterest(Double settleInterest) {
+    public void setSettleInterest(Long settleInterest) {
         this.settleInterest = settleInterest;
     }
